# swagger.yaml
basePath: /
consumes:
- application/json
definitions:
...
  GetDetailProductsResponse:
    properties:
      description:
        type: string
        x-go-name: Description
      image:
        type: string
        x-go-name: Image
      name:
        type: string
        x-go-name: Name
      url:
        type: string
        x-go-name: Url
    type: object
    x-go-package: _/Users/waresix/Code/sut-gateway-go/domain/order/response
...
  Response:
    properties:
      data:
        type: object
        x-go-name: Data
      message:
        type: string
        x-go-name: Message
      status:
        type: string
        x-go-name: Status
      statusCode:
        format: int64
        type: integer
        x-go-name: StatusCode
      timestamp:
        type: string
        x-go-name: Timestamp
    type: object
    x-go-package: _/Users/waresix/Code/sut-gateway-go/helpers/http_response
...
info:
  description: Documentation for sut api gateway
  title: SUT-gateway-go.
  version: 0.1.0
paths:
...
  /order/bulk:
    post:
      description: Create bulk order
      operationId: CreateOrderBulkHeader
      parameters:
      - example: JWT eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          $ref: '#/responses/swaggerResponse'
        "400":
          $ref: '#/responses/swaggerResponse'
      tags:
      - Order
  /order/detail-product:
    get:
      description: Get product detail
      operationId: GetDetailProductsQuery
      parameters:
      - in: query
        name: Url
        type: string
      responses:
        "200":
          $ref: '#/responses/GetDetailProductsResponse'
        "400":
          $ref: '#/responses/swaggerResponse'
      tags:
      - Order
  /order/products-by-keyword:
    get:
      description: Get products from bukalapak
      operationId: GetProductsToOrderByKeywordQuery
      parameters:
      - in: query
        name: Keyword
        type: string
      responses:
        "200":
          $ref: '#/responses/GetProductsToOrderResponse'
        "400":
          $ref: '#/responses/swaggerResponse'
      tags:
      - Order
produces:
- application/json
responses:
...
  GetDetailProductsResponse:
    description: ""
    schema:
      $ref: '#/definitions/GetDetailProductsResponse'
...
  GetProductsToOrderResponse:
    description: ""
    schema:
      $ref: '#/definitions/GetDetailProductsResponse'
...
  swaggerResponse:
    description: ""
    schema:
      $ref: '#/definitions/Response'
schemes:
- http
securityDefinitions:
  basic:
    type: basic
swagger: "2.0"